lambda/tools/index.py
import json
import boto3
import os

def handler(event, context):
    """
    Lambda function serving as a tool for the AI agent.
    This can be extended to handle various tool functionalities.
    """
    try:
        # Initialize Bedrock client
        bedrock = boto3.client('bedrock-runtime')
        
        # Process the event based on the tool type
        tool_type = event.get('tool_type')
        
        if tool_type == 'external_api':
            # Handle external API requests
            return handle_external_api(event)
        elif tool_type == 'knowledge_base':
            # Handle knowledge base queries
            return handle_knowledge_base(event, bedrock)
        else:
            raise ValueError(f"Unsupported tool type: {tool_type}")
            
    except Exception as e:
        return {
            'statusCode': 500,
            'body': json.dumps({
                'error': str(e)
            })
        }

def handle_external_api(event):
    # Implement external API integration
    pass

def handle_knowledge_base(event, bedrock_client):
    # Implement Bedrock knowledge base integration
    pass