requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
pytest>=6.0
aws-cdk.aws-codecommit>=2.0.0
aws-cdk.aws-codepipeline>=2.0.0
aws-cdk.aws-codebuild>=2.0.0
aws-cdk.aws-lambda>=2.0.0
aws-cdk.aws-s3>=2.0.0
aws-cdk.aws-batch>=2.0.0
aws-cdk.aws-bedrock>=2.0.0
aws-cdk.aws-cloudwatch>=2.0.0